cherry-picked Marc's assertEqual code
