+++ Tests/SwiftFusionTests/Core/MatrixTests.swift
@@ -3,27 +3,6 @@ import XCTest
 import TensorFlow
 import SwiftFusion
 
-/// Asserts that `x` and `y` have the same shape and that their values have absolute difference
-/// less than `accuracy`.
-func assertEqual<T: TensorFlowFloatingPoint>(
-  _ x: Tensor<T>, _ y: Tensor<T>, accuracy: T, file: StaticString = #file, line: UInt = #line
-) {
-  guard x.shape == y.shape else {
-    XCTFail(
-      "shape mismatch: \(x.shape) is not equal to \(y.shape)",
-      file: file,
-      line: line
-    )
-    return
-  }
-  XCTAssert(
-    abs(x - y).max().scalarized() < accuracy,
-    "value mismatch:\n\(x)\nis not equal to\n\(y)\nwith accuracy \(accuracy)",
-    file: file,
-    line: line
-  )
-}
-
 class MatrixTests: XCTestCase {
     static var allTests = [
         ("test_concat", testConcat),
